requirements.txt
django==2.0.7
flake8
tox
djangorestframework==3.7.7
django_extensions
markdown
django-filter
certifi==2019.3.9
chardet==3.0.4
idna==2.8
nba-api==1.1.4
numpy==1.16.2
pandas==0.24.2
python-dateutil==2.8.0
pytz==2018.9
requests==2.21.0
six==1.12.0
urllib3==1.24.1
beautifulsoup4==4.7.1
bs4==0.0.1
html5lib==1.0.1
six==1.12.0
soupsieve==1.9
webencodings==0.5.1
celery==4.3.0
djangorestframework-jwt
psycopg2
